Fixed bug where sleeves of <$1.00 value weren't showing up in remainder total. This was caused by an unintended rounding from int conversion.

+++ src/Currency/Coin.java
@@ -11,7 +11,7 @@ public class Coin {
         this.name = name;
         this.value = val;
         this.roll_limit = roll_limit;
-        this.roll_value = val * roll_limit;
+        this.roll_value = (float)(val * roll_limit);
     }
 
     public String getName() {
